add offset, limit to dump result, move dump code to new file
